// browser/src/libs/keyboard/layouts.generated.ts
export const GENERATED_LAYOUTS: Record<string, { name: string; map: LayoutMap }> = {
  // Albanian
  'al': {
    name: 'Albanian',
    map: {
// ...
      '-': { code: 0x2d },
// ...
      '.': { code: 0x37 },
      ':': { code: 0x37, shift: true },
      '>': { code: 0x37, altGr: true },
      '/': { code: 0x38 },
      '?': { code: 0x38, shift: true },
      ' ': { code: 0x2c },
      '\t': { code: 0x2b },
      '\n': { code: 0x28 },
      '\r': { code: 0x28 },
    },
  },
// ...
};